# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tratamiento-insumo": {
            "get": {
                "tags": [
                    "TratamientoInsumo"
                ],
                "summary": "Obtener todas las relaciones tratamiento-insumo",
                "operationId": "6842b004a6b4d249771affc5f240f6fd",
                "responses": {
                    "200": {
                        "description": "Relaciones encontradas correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TratamientoInsumo"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron relaciones"
                    }
                }
            },
            "post": {
                "tags": [
                    "TratamientoInsumo"
                ],
                "summary": "Crear una relación tratamiento-insumo",
                "operationId": "a3fe096c843cbcade2a96812ec20b143",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_tratamiento",
                                    "id_insumo"
                                ],
                                "properties": {
                                    "id_tratamiento": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_insumo": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Relación creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TratamientoInsumo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error en la validación de datos"
                    }
                }
            }
        },
        "/api/tratamiento-insumo/{id}": {
            "get": {
                "tags": [
                    "TratamientoInsumo"
                ],
                "summary": "Obtener una relación tratamiento-insumo por ID",
                "operationId": "0e4d2f483e6f4b531843e60c277190ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación encontrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TratamientoInsumo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "TratamientoInsumo"
                ],
                "summary": "Actualizar una relación tratamiento-insumo",
                "operationId": "b75e72476fa4dace896321d287ef059b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_tratamiento": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_insumo": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Relación actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TratamientoInsumo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    },
                    "422": {
                        "description": "Error en la validación de datos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TratamientoInsumo"
                ],
                "summary": "Eliminar una relación tratamiento-insumo por ID",
                "operationId": "b9626d688044e6455f903692f9808c45",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la relación a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Relación eliminada correctamente"
                    },
                    "404": {
                        "description": "Relación no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "TratamientoInsumo": {
                "title": "TratamientoInsumo",
                "required": [
                    "id",
                    "id_tratamiento",
                    "id_insumo"
                ],
                "properties": {
                    "id": {
                        "description": "ID de la relación",
                        "type": "integer",
                        "example": 1
                    },
                    "id_tratamiento": {
                        "description": "ID del tratamiento",
                        "type": "integer",
                        "example": 1
                    },
                    "id_insumo": {
                        "description": "ID del insumo",
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-04T15:00:00Z"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-07-04T15:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}